+++ src/bindings/point3d_bindings.cpp
@@ -10,6 +10,7 @@ void Point3dBindings(nb::module_& m) {
     nb::class_<ON_3dPoint>(m, "Point3d")
         /*dunder methods*/
         .def(nb::init<double, double, double>())
+        .def("__add__", &ON_3dPoint::operator+=)
         .def("__eq__", &ON_3dPoint::operator==)
         .def("__mul__", &ON_3dPoint::operator*=)
         .def("__ne__", &ON_3dPoint::operator!=)